requirements_pip.txt
antlr4-python3-runtime==4.9.3
asrl-pylgmath==1.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
cython==3.0.8
fonttools==4.45.1
gym==0.26.2
gym-notices==0.0.8
hydra-core==1.3.2
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
markupsafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
omegaconf==2.3.0
packaging==23.2
pandas==2.1.4
pillow==10.1.0
psutil==5.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
shapely==2.0.2
threadpoolctl==3.2.0
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.4
urllib3==2.1.0
zipp==3.17.0